Terminado

+++ backend/controllers/tareasControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')//para manejar errores en l
 const Tarea = require('../models/tareasModel')
 
 const getTareas = asyncHandler( async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
     res.status(200).json(tareas)
 })
 
@@ -12,7 +12,8 @@ const setTarea = asyncHandler( async (req, res) => {
         throw new Error('Por favor teclea una Tarea')
     }
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
     })
     res.status(201).json(tarea)
 })
@@ -26,7 +27,14 @@ const updtaTarea = asyncHandler( async (req, res) => {
     throw new Error ('La Tarea no fue encontrada')
   }
 
-  const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+  //verificamos que la tarea pertenesca al usuario logeado
+
+  if(tarea.user.toString() !== req.user.id){
+    res.status(401)
+    throw new Error ('Usuario no autorizado')
+  } else {
+    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+  }
 
     res.status(200).json(updatedTarea)
 })
@@ -40,8 +48,15 @@ const deleteTarea = asyncHandler( async (req, res) => {
       throw new Error ('La Tarea no fue encontrada')
     }
 
+
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error ('Usuario no autorizado')
+      } else {
+        tarea.deleteOne()
+      }
+
     //const deletedTarea = await Tarea.findByIdAndDelete(req.params.id) No deria necesario pues ya se encontro 
-    tarea.deleteOne()
 
     res.status(200).json({id: tarea.id})
 })
